Decreased the Width of the Attribute and Song containers
This prevents the tooltip from appearing when the user is far away from an attribute label. Although the width for song container has been decreased, the tooltip will appear even when the user is still pretty far away. I could not figure out a work around for this because the width of each song varies significantly.